pyproject.toml
[project]
name = "agentic-rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.5.5",
    "docx2txt>=0.9",
    "google-generativeai>=0.8.5",
    "langchain-text-splitters>=0.3.8",
    "openai-agents[litellm]>=0.0.14",
    "pillow>=11.2.1",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-magic-bin>=0.4.14",
    "qdrant-client>=1.14.2",
    "textract>=1.6.5",
]
